+++ src/python/arcor2_mocks/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## [0.15.0] - WIP
+
+### Changed
+- Updated to match API of Project service 0.8.0.
+
 ## [0.14.0] - 2021-05-21
 
 ### Changed